+++ functions.php
@@ -23,6 +23,7 @@ require_once get_theme_file_path( 'inc/schema.php' );
 require_once get_theme_file_path( 'inc/svg.php' );
 require_once get_theme_file_path( 'inc/woocommerce.php' );
 require_once get_theme_file_path( 'inc/performance.php' );
+require_once get_theme_file_path( 'inc/blocks.php' );
 
 if ( ! function_exists( 'agramlabs_setup' ) ) :
 	/**